+++ beacon/sync/src/testFixtures/java/tech/pegasys/teku/beacon/sync/forward/multipeer/batches/StubBatchFactory.java
@@ -29,26 +29,26 @@ import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.networking.eth2.peers.StubSyncSource;
 import tech.pegasys.teku.networking.eth2.peers.SyncSource;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
-import tech.pegasys.teku.statetransition.blobs.BlobsSidecarManager;
+import tech.pegasys.teku.statetransition.blobs.BlobSidecarManager;
 
 public class StubBatchFactory extends BatchFactory implements Iterable<Batch> {
   private final List<Batch> batches = new ArrayList<>();
   private final Map<Batch, BatchSupport> batchSupports = new HashMap<>();
 
   private final EventThread eventThread;
-  private final BlobsSidecarManager blobsSidecarManager;
+  private final BlobSidecarManager blobSidecarManager;
   private final boolean enforceEventThread;
 
   public StubBatchFactory(final EventThread eventThread, final boolean enforceEventThread) {
-    this(eventThread, BlobsSidecarManager.NOOP, enforceEventThread);
+    this(eventThread, BlobSidecarManager.NOOP, enforceEventThread);
   }
 
   public StubBatchFactory(
       final EventThread eventThread,
-      final BlobsSidecarManager blobsSidecarManager,
+      final BlobSidecarManager blobSidecarManager,
       final boolean enforceEventThread) {
-    super(eventThread, blobsSidecarManager, null);
-    this.blobsSidecarManager = blobsSidecarManager;
+    super(eventThread, blobSidecarManager, null);
+    this.blobSidecarManager = blobSidecarManager;
     this.eventThread = eventThread;
     this.enforceEventThread = enforceEventThread;
   }
@@ -100,7 +100,7 @@ public class StubBatchFactory extends BatchFactory implements Iterable<Batch> {
   @Override
   public Batch createBatch(final TargetChain chain, final UInt64 start, final UInt64 count) {
     final BatchSupport support =
-        new BatchSupport(eventThread, blobsSidecarManager, chain, start, count);
+        new BatchSupport(eventThread, blobSidecarManager, chain, start, count);
     batches.add(support.batch);
     // Can look up batch support by either the wrapped or unwrapped batch
     batchSupports.put(support.batch, support);
@@ -129,12 +129,11 @@ public class StubBatchFactory extends BatchFactory implements Iterable<Batch> {
 
     public BatchSupport(
         final EventThread eventThread,
-        final BlobsSidecarManager blobsSidecarManager,
+        final BlobSidecarManager blobSidecarManager,
         final TargetChain chain,
         final UInt64 start,
         final UInt64 count) {
-      batch =
-          new SyncSourceBatch(eventThread, blobsSidecarManager, this, this, chain, start, count);
+      batch = new SyncSourceBatch(eventThread, blobSidecarManager, this, this, chain, start, count);
       eventThreadOnlyBatch = new EventThreadOnlyBatch(eventThread, batch);
     }