+++ validator/api/src/main/java/tech/pegasys/teku/validator/api/AttesterDuties.java
@@ -20,27 +20,20 @@ import org.apache.tuweni.bytes.Bytes32;
 
 public class AttesterDuties {
   private final boolean executionOptimistic;
-  private final boolean finalized;
   private final Bytes32 dependentRoot;
   private final List<AttesterDuty> duties;
 
   public AttesterDuties(
       final boolean executionOptimistic,
-      final boolean finalized,
       final Bytes32 dependentRoot,
       final List<AttesterDuty> duties) {
     this.executionOptimistic = executionOptimistic;
-    this.finalized = finalized;
     this.dependentRoot = dependentRoot;
     this.duties = duties;
   }
 
   public boolean isExecutionOptimistic() {
     return executionOptimistic;
-  }
-
-  public boolean isFinalized() {
-    return finalized;
   }
 
   public Bytes32 getDependentRoot() {
@@ -61,21 +54,19 @@ public class AttesterDuties {
     }
     final AttesterDuties that = (AttesterDuties) o;
     return executionOptimistic == that.executionOptimistic
-        && finalized == that.finalized
         && Objects.equals(dependentRoot, that.dependentRoot)
         && Objects.equals(duties, that.duties);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(executionOptimistic, finalized, dependentRoot, duties);
+    return Objects.hash(executionOptimistic, dependentRoot, duties);
   }
 
   @Override
   public String toString() {
     return MoreObjects.toStringHelper(this)
         .add("executionOptimistic", executionOptimistic)
-        .add("finalized", finalized)
         .add("dependentRoot", dependentRoot)
         .add("duties", duties)
         .toString();