+++ validator/remote/src/test/java/tech/pegasys/teku/validator/remote/RemoteValidatorApiHandlerTest.java
@@ -67,6 +67,7 @@ import tech.pegasys.teku.spec.TestSpecFactory;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlock;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.spec.datastructures.blocks.versions.deneb.BlindedBlockContents;
+import tech.pegasys.teku.spec.datastructures.blocks.versions.deneb.BlockContents;
 import tech.pegasys.teku.spec.datastructures.builder.SignedValidatorRegistration;
 import tech.pegasys.teku.spec.datastructures.genesis.GenesisData;
 import tech.pegasys.teku.spec.datastructures.operations.AggregateAndProof;
@@ -240,7 +241,7 @@ class RemoteValidatorApiHandlerTest {
         .thenReturn(
             Optional.of(
                 new PostAttesterDutiesResponse(
-                    dataStructureUtil.randomBytes32(), Collections.emptyList(), false, false)));
+                    dataStructureUtil.randomBytes32(), Collections.emptyList(), false)));
 
     SafeFuture<Optional<AttesterDuties>> future =
         apiHandler.getAttestationDuties(UInt64.ONE, IntList.of(1234));
@@ -279,10 +280,7 @@ class RemoteValidatorApiHandlerTest {
         .thenReturn(
             Optional.of(
                 new PostAttesterDutiesResponse(
-                    dataStructureUtil.randomBytes32(),
-                    List.of(schemaValidatorDuties),
-                    false,
-                    false)));
+                    dataStructureUtil.randomBytes32(), List.of(schemaValidatorDuties), false)));
 
     SafeFuture<Optional<AttesterDuties>> future =
         apiHandler.getAttestationDuties(UInt64.ONE, IntList.of(validatorIndex));
@@ -411,7 +409,7 @@ class RemoteValidatorApiHandlerTest {
   }
 
   @Test
-  public void createUnsignedBlindedBlockContents_WhenFound_ReturnsBlockContents() {
+  public void createUnsignedBlindedBlockContents_WhenFound_ReturnsBlindedBlockContents() {
     final Spec denebSpec = TestSpecFactory.createMinimalDeneb();
     final DataStructureUtil denebDataStructureUtil = new DataStructureUtil(denebSpec);
     final BeaconBlock beaconBlock = denebDataStructureUtil.randomBeaconBlock(UInt64.ONE);
@@ -428,6 +426,25 @@ class RemoteValidatorApiHandlerTest {
         apiHandler.createUnsignedBlindedBlockContents(UInt64.ONE, blsSignature, graffiti);
 
     assertThatSszData(unwrapToValue(future)).isEqualByAllMeansTo(blindedBlockContents);
+  }
+
+  @Test
+  public void createUnsignedBlockContents_WhenFound_ReturnsBlockContents() {
+    final Spec denebSpec = TestSpecFactory.createMinimalDeneb();
+    final DataStructureUtil denebDataStructureUtil = new DataStructureUtil(denebSpec);
+    final BeaconBlock beaconBlock = denebDataStructureUtil.randomBeaconBlock(UInt64.ONE);
+    final BlockContents blockContents = denebDataStructureUtil.randomBlockContents(UInt64.ONE);
+    final BLSSignature blsSignature = denebDataStructureUtil.randomSignature();
+    final Optional<Bytes32> graffiti = Optional.of(Bytes32.random());
+
+    when(typeDefClient.createUnsignedBlockContents(
+            eq(beaconBlock.getSlot()), eq(blsSignature), eq(graffiti)))
+        .thenReturn(Optional.of(blockContents));
+
+    SafeFuture<Optional<BlockContents>> future =
+        apiHandler.createUnsignedBlockContents(UInt64.ONE, blsSignature, graffiti);
+
+    assertThatSszData(unwrapToValue(future)).isEqualByAllMeansTo(blockContents);
   }
 
   @Test